Update focus indicators
Removes flying focus CSS and JS so we have less stuff that can go wrong.
Adds a new JS to get focus to show up in a nicer way for links that
contain other elements like headings and images, as well as the table
headers which aren't semantically correct yet.